Enable api_v3_syntaxConformaceTest::testInvalidID_delete (#9068)
* Enable api_v3_SyntaxConformanceTest delete with wrong id

* Remove unneded if and remove debugging

* Remove debugging

* Tidy up fix for CustomValue tear down

+++ CRM/Event/BAO/ParticipantStatusType.php
@@ -80,7 +80,9 @@ class CRM_Event_BAO_ParticipantStatusType extends CRM_Event_DAO_ParticipantStatu
 
     $dao = new CRM_Event_DAO_ParticipantStatusType();
     $dao->id = $id;
-    $dao->find(TRUE);
+    if (!$dao->find()) {
+      return FALSE;
+    }
     $dao->delete();
     return TRUE;
   }
